+++ BO360/BO360/BOStationaryParticle.cpp
@@ -46,7 +46,7 @@ void BOStationaryParticle::Update(double p_deltaTime)
     if (m_timeLeftS <= m_startFadeTime)
     {
         double fadePercentage = m_timeLeftS / m_startFadeTime;
-        m_opacity = (255 * fadePercentage);
+        m_opacity = (Uint8)(255 * fadePercentage);
         
     }
